+++ packages/sdk/src/ui/with_hooks.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import {spawnError} from '../error_utils';
 
 /**
- * @internal
+ * @hidden
  * A higher-order component for working with React hooks in class-based components. It takes a React
  * component and wraps it, injecting values from hooks as additional props. `withHooks` uses
  * [`React.forwardRef`](https://reactjs.org/docs/forwarding-refs.html) to make sure that you can